+++ src/main/java/ssuPlector/ai/openAI/service/ChatGptService.java
@@ -1,5 +1,8 @@
 package ssuPlector.ai.openAI.service;
 
 public interface ChatGptService {
+
+    String summarizeText(String text);
+
     String recommendMeetingToDo(String query);
 }
