--- src/main/java/ssuPlector/ai/openAi/dto/ChatGptResponse.java
@@ -1,23 +1,0 @@
-package ssuPlector.ai.openAi.dto;
-
-import java.util.List;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ChatGptResponse {
-
-    private List<Choice> choices;
-
-    @Getter
-    @Setter
-    @AllArgsConstructor
-    @NoArgsConstructor
-    public static class Choice {
-        private int index;
-        private Message Message;
-    }
-}
