--- src/main/java/ssuPlector/ai/openAi/dto/Message.java
@@ -1,12 +1,0 @@
-package ssuPlector.ai.openAi.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class Message {
-    private String role;
-    private String content;
-}
