--- src/main/java/ssuPlector/ai/openAi/service/ChatGptService.java
@@ -1,6 +1,0 @@
-package ssuPlector.ai.openAi.service;
-
-public interface ChatGptService {
-
-    String summarizeText(String text);
-}
