+++ src/main/java/ssuPlector/security/SecurityConfig.java
@@ -68,7 +68,7 @@ public class SecurityConfig {
                                         "/api/projects/list",
                                         "/api/projects",
                                         "/api/developers/dummy",
-                                        "/api/pm/assistant/summary")
+                                        "/api/assistant/pm/summary",
                                         "/api/assistant/pm/meeting")
                                 .permitAll()
                                 .anyRequest()
