--- src/main/java/ssuPlector/service/ai/PmService.java
@@ -1,8 +1,0 @@
-package ssuPlector.service.ai;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public interface PmService {
-
-    String summarize(MultipartFile file);
-}
